User Edit Screen & Get User Details

+++ frontend/src/screens/UserListScreen.js
@@ -24,11 +24,10 @@ const UserListScreen = ({ history }) => {
     } else {
       history.push("/login")
     }
-  }, [dispatch, history, successDelete])
+  }, [dispatch, history, successDelete, userInfo])
 
   const deleteHandler = (id) => {
-    if(window.confirm('Are you sure?'))
-    dispatch(deleteUser(id))
+    if (window.confirm("Are you sure?")) dispatch(deleteUser(id))
   }
 
   return (
